fix: ResultPageをサーバー/クライアントコンポーネントに分離（async Client Component エラー修正）
